requirements.txt
openai
reportlab
moviepy
elevenlabs
rich
python-dotenv
tiktoken
pydantic
